Split assistant.js
