+++ html/index.html
@@ -503,6 +503,10 @@
     <script src="js/routes/config.js?v=VERSION_PLACEHOLDER"></script>
     <script src="js/routes/downloads.js?v=VERSION_PLACEHOLDER"></script>
     <script src="js/routes/home.js?v=VERSION_PLACEHOLDER"></script>
+    <script src="js/routes/assistant.sessions.js?v=VERSION_PLACEHOLDER"></script>
+    <script src="js/routes/assistant.streaming.js?v=VERSION_PLACEHOLDER"></script>
+    <script src="js/routes/assistant.tools.js?v=VERSION_PLACEHOLDER"></script>
+    <script src="js/routes/assistant.utils.js?v=VERSION_PLACEHOLDER"></script>
     <script src="js/routes/assistant.js?v=VERSION_PLACEHOLDER"></script>
     <script src="js/routes/hunt.js?v=VERSION_PLACEHOLDER"></script>
     <script src="js/routes/jobs.js?v=VERSION_PLACEHOLDER"></script>
